Remove esc global keybinding
Moved it to a textarea element keybinding so the editor has more control
over what is happening. With this change came a few other changes
because I needed to make sure the stmt would go back to normal mode on
exit and that the cursor and sql values would save as well.

This also had one cleanup of syntax where I deconstruct the props into
individual variables so we can see what the component requires.